feat(expand-collapse): Enable conditional rendering of panels (#447)
* feat(expand-collapse): Enable conditional rendering of panels

Allows developers to use common react patterns (ternary operators or the use of &&) to conditionally
render panels inside the expand-collapse component

444

* chore(expand-collapse): Address code-review feedback

Use the panelId for the key, should be safe as this is required and unique in the expand. Streamline
the test for conditional render and update tests to pass ids to avoide duplicate key warning.