Update SpinnerSvg.jsx
That piece of CJS breaks vite(esbuild) production build. I tried fixing it the way as in this PR in the build file of your package and it looked like working. Only thing is left to check the build by rollup whether that works well.
And also to the commented TODO issue in rollup: last comment in that issue states that it's not an issue anymore. Anyone can run tests on this?